+++ bundles/lupapiste/bundle/lupapiste-myplaces2/plugin/DrawPlugin.js
@@ -38,7 +38,7 @@ Oskari.clazz.define('Oskari.lupapiste.bundle.myplaces2.plugin.DrawPlugin', funct
         }
         else {
 	        // remove possible old drawing
-	        this.drawLayer.removeAllFeatures();
+	        this.drawLayer.destroyFeatures();
         	
 	        if(params.geometry) {
 	            // sent existing geometry == edit mode
@@ -66,7 +66,7 @@ Oskari.clazz.define('Oskari.lupapiste.bundle.myplaces2.plugin.DrawPlugin', funct
         // disable all draw controls
         this.toggleControl();
         // clear drawing
-        this.drawLayer.removeAllFeatures();
+        this.drawLayer.destroyFeatures();
     },
     
     forceFinishDraw : function() {
@@ -165,11 +165,20 @@ Oskari.clazz.define('Oskari.lupapiste.bundle.myplaces2.plugin.DrawPlugin', funct
         
         // doesn't really need to be in array, but lets keep it for future development
         this.modifyControls = {
-        	select : new OpenLayers.Control.SelectFeature(me.drawLayer),
-        	modify : new OpenLayers.Control.ModifyFeature(me.drawLayer, {
+            modify : new OpenLayers.Control.ModifyFeature(me.drawLayer, {
                 standalone: true
             })
         };
+        this.modifyControls.select = new OpenLayers.Control.SelectFeature(me.drawLayer, {
+            geometryTypes: this.modifyControls.modify.geometryTypes,
+            clickout: this.modifyControls.modify.clickout,
+            toggle: this.modifyControls.modify.toggle,
+            onBeforeSelect: this.modifyControls.modify.beforeSelectFeature,
+            onSelect: this.modifyControls.modify.selectFeature,
+            onUnselect: this.modifyControls.modify.unselectFeature,
+            scope: this.modifyControls.modify
+        });
+        
         this._map.addLayers([me.drawLayer]);
         for(var key in this.drawControls) {
             this._map.addControl(this.drawControls[key]);