+++ ReleaseNotes.md
@@ -15,6 +15,8 @@ Environment specific localized values (URLs) have been move to bundle configurat
 
 Added new method to create maplayer domain objects based on type: createLayerTypeInstance(type). This is a preferred way to create layer domain classes instead of Oskari.clazz.create() if you need to create one manually.
 
+Added new method to find all layers of given type: getLayersOfType(type). For example get all wfs layers by calling getLayersOfType('wfs').
+
 ### mapmodule-plugin/layers/backgroundlayerselector
 
 New plugin for selecting a background layer from a preset list. See the bundle documentation for more information.
@@ -85,6 +87,8 @@ User interface bug fixes.
 
 Sorting is now disabled when clicking the header menu buttons of an indicator in the grid.
 The classification now shows distinct class ranges thanks to the geostats library update.
+
+Bundle now has a tile for easier access to statistics. Statistics layer to use can be configured with defaultLayerId : [layer id]
 
 ### libraries/geostats